requirements.txt
torch>=2.0.0
torchvision>=0.15.0
clip @ git+https://github.com/openai/CLIP.git
scikit-learn>=1.3.0
numpy>=1.24.0
Pillow>=10.0.0
tqdm>=4.65.0
PyYAML>=6.0
